refactor: extract I18nString with generic
