+++ packages/backend/tools/limitlesstcg-game-data/src/cards/adapter/limitlesstcg-pokemon/models/TcgPokemonParsedCard.ts
@@ -1,3 +1,4 @@
+import { I18nString } from '../../../../common/domain/models/I18nString';
 import { TcgPokemonLanguage } from '../../../../languages/adapter/limitlesstcg-pokemon/models/TcgPokemonLanguage';
 import { TcgPokemonCardSubtype } from './TcgPokemonCardSubtype';
 import { TcgPokemonCardType } from './TcgPokemonCardType';
@@ -5,19 +6,15 @@ import { TcgPokemonRarity } from './TcgPokemonRarity';
 import { TcgPokemonStage } from './TcgPokemonStage';
 import { TcgPokemonType } from './TcgPokemonType';
 
-export type I18nString = {
-  [TLang in TcgPokemonLanguage]?: string;
-};
-
 export interface BaseTcgPokemonParsedCard<TType extends TcgPokemonCardType> {
   cardType: TType;
   imageUrls: {
-    xs: I18nString;
-    sm: I18nString;
-    lg: I18nString;
+    xs: I18nString<TcgPokemonLanguage>;
+    sm: I18nString<TcgPokemonLanguage>;
+    lg: I18nString<TcgPokemonLanguage>;
   };
   languages: TcgPokemonLanguage[];
-  name: I18nString;
+  name: I18nString<TcgPokemonLanguage>;
   number: string;
   rarity?: TcgPokemonRarity | undefined;
   regulation?: string | undefined;
@@ -40,20 +37,20 @@ export enum TcgPokemonParsedPokemonCardAttackCostType {
 export interface TcgPokemonParsedPokemonCardAttack {
   cost?: TcgPokemonParsedPokemonCardAttackCostType[] | undefined;
   dmg?: string | undefined;
-  effect: I18nString;
-  name: I18nString;
+  effect: I18nString<TcgPokemonLanguage>;
+  name: I18nString<TcgPokemonLanguage>;
 }
 
 export interface TcgPokemonParsedEnergyCard
   extends BaseTcgPokemonParsedCard<TcgPokemonCardType.energy> {
-  effect: I18nString;
+  effect: I18nString<TcgPokemonLanguage>;
   type: TcgPokemonCardSubtype.basic | TcgPokemonCardSubtype.special;
 }
 
 export interface TcgPokemonParsedPokemonCard
   extends BaseTcgPokemonParsedCard<TcgPokemonCardType.pokemon> {
-  abilityName?: I18nString | undefined;
-  abilityEffect?: I18nString | undefined;
+  abilityName?: I18nString<TcgPokemonLanguage> | undefined;
+  abilityEffect?: I18nString<TcgPokemonLanguage> | undefined;
   attacks: TcgPokemonParsedPokemonCardAttack[];
   hp: number;
   resistance: TcgPokemonType[];
@@ -67,7 +64,7 @@ export interface TcgPokemonParsedPokemonCard
 
 export interface TcgPokemonParsedTrainerCard
   extends BaseTcgPokemonParsedCard<TcgPokemonCardType.trainer> {
-  effect: I18nString;
+  effect: I18nString<TcgPokemonLanguage>;
   type:
     | TcgPokemonCardSubtype.item
     | TcgPokemonCardSubtype.itemTm
