+++ packages/backend/tools/limitlesstcg-game-data/src/common/domain/models/I18nString.ts
@@ -1,0 +1,3 @@
+export type I18nString<TLanguage extends string = string> = {
+  [TLang in TLanguage]?: string;
+};
