+++ packages/backend/tools/limitlesstcg-game-data/src/index.ts
@@ -1,6 +1,5 @@
 export { getParsedSetCards } from './cards/adapter/limitlesstcg-pokemon/calculations/getParsedSetCards';
 export type {
-  I18nString,
   TcgPokemonParsedCard,
   TcgPokemonParsedEnergyCard,
   TcgPokemonParsedPokemonCard,
@@ -11,6 +10,8 @@ export type {
 export { TcgPokemonCardType } from './cards/adapter/limitlesstcg-pokemon/models/TcgPokemonCardType';
 export { TcgPokemonRarity } from './cards/adapter/limitlesstcg-pokemon/models/TcgPokemonRarity';
 export { TcgPokemonCardSubtype } from './cards/adapter/limitlesstcg-pokemon/models/TcgPokemonCardSubtype';
+export type { I18nString } from './common/domain/models/I18nString';
+export { TcgPokemonLanguage } from './languages/adapter/limitlesstcg-pokemon/models/TcgPokemonLanguage';
 export { TcgPokemonRegion } from './regions/adapter/limitlesstcg-pokemon/models/TcgPokemonRegion';
 export { getParsedSets } from './sets/adapter/limitlesstcg-pokemon/calculations/getParsedSets';
 export type { TcgPokemonParsedSet } from './sets/adapter/limitlesstcg-pokemon/models/TcgPokemonParsedSet';
