+++ packages/backend/tools/limitlesstcg-game-data/src/sets/adapter/limitlesstcg-pokemon/models/TcgPokemonParsedSet.ts
@@ -1,8 +1,9 @@
-import { I18nString } from '../../../../cards/adapter/limitlesstcg-pokemon/models/TcgPokemonParsedCard';
+import { I18nString } from '../../../../common/domain/models/I18nString';
+import { TcgPokemonLanguage } from '../../../../languages/adapter/limitlesstcg-pokemon/models/TcgPokemonLanguage';
 import { TcgPokemonRegion } from '../../../../regions/adapter/limitlesstcg-pokemon/models/TcgPokemonRegion';
 
 export interface TcgPokemonParsedSet {
   code: string;
-  name: I18nString;
+  name: I18nString<TcgPokemonLanguage>;
   region: TcgPokemonRegion;
 }
